[#56] chore: 파일을 분리함
